# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Gereh Public API
  version: 0.1.0
  description: |
    Public REST API exposed by the Gereh API BFF.
servers:
  - url: /
...
paths:
  /v1/status:
    get:
      operationId: getSystemStatus
      summary: Get API status
      description: Returns API BFF health and version information.
      tags:
        - System
      responses:
        "200":
          description: The API is available.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatus"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/tenants/{tenantId}/goals:
    post:
      operationId: createGoal
      summary: Create a goal
      description: Creates a goal under a company.
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Goal created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateGoalResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: listGoals
      summary: List goals
      description: Lists goals, optionally filtered by company.
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: company_id
          in: query
          schema:
            type: string
            format: uuid
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_token
          in: query
          schema:
            type: string
        - name: include_archived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Goal page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListGoalsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/tenants/{tenantId}/goals/{goalId}:
    get:
      operationId: getGoal
      summary: Get a goal
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Goal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetGoalResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGoal
      summary: Update a goal
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalRequest"
      responses:
        "200":
          description: Updated goal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateGoalResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenantId}/goals/{goalId}/status:
    post:
      operationId: changeGoalStatus
      summary: Change a goal status
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeStatusRequest"
      responses:
        "200":
          description: Updated goal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeGoalStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenantId}/goals/{goalId}/projects:
    post:
      operationId: createProject
      summary: Create a project
      description: Creates a project under a goal.
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: listProjects
      summary: List projects
      description: Lists projects, optionally filtered by company.
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: company_id
          in: query
          schema:
            type: string
            format: uuid
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_token
          in: query
          schema:
            type: string
        - name: include_archived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Project page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProjectsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/tenants/{tenantId}/projects/{projectId}:
    get:
      operationId: getProject
      summary: Get a project
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProjectResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      summary: Update a project
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenantId}/projects/{projectId}/status:
    post:
      operationId: changeProjectStatus
      summary: Change a project status
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeStatusRequest"
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeProjectStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  headers:
    RequestId:
      description: Identifier used to correlate logs and traces.
      schema:
        type: string
        minLength: 1
  schemas:
...
    SystemStatus:
      type: object
      additionalProperties: false
      required:
        - status
        - service
        - version
      properties:
        status:
          type: string
          const: ok
        service:
          type: string
          minLength: 1
          examples:
            - api-bff
        version:
          type: string
          minLength: 1
          examples:
            - dev
    CreateGoalRequest:
      type: object
      additionalProperties: false
      required:
        - companyId
        - title
      properties:
        companyId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 4000
    CreateGoalResponse:
      type: object
      additionalProperties: false
      required:
        - goal
      properties:
        goal:
          $ref: "#/components/schemas/Goal"
    GetGoalResponse:
      type: object
      additionalProperties: false
      required:
        - goal
      properties:
        goal:
          $ref: "#/components/schemas/Goal"
    ListGoalsResponse:
      type: object
      additionalProperties: false
      properties:
        goals:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
        nextPageToken:
          type: string
    UpdateGoalRequest:
      type: object
      additionalProperties: false
      required:
        - expectedVersion
      properties:
        expectedVersion:
          type: integer
          format: int64
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 4000
    UpdateGoalResponse:
      type: object
      additionalProperties: false
      required:
        - goal
      properties:
        goal:
          $ref: "#/components/schemas/Goal"
    ChangeStatusRequest:
      type: object
      additionalProperties: false
      required:
        - expectedVersion
        - status
      properties:
        expectedVersion:
          type: integer
          format: int64
        status:
          type: integer
          minimum: 1
    ChangeGoalStatusResponse:
      type: object
      additionalProperties: false
      required:
        - goal
      properties:
        goal:
          $ref: "#/components/schemas/Goal"
    Goal:
      type: object
      additionalProperties: false
      required:
        - tenantId
        - companyId
        - goalId
        - title
        - status
        - version
      properties:
        tenantId:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - GOAL_STATUS_ACTIVE
            - GOAL_STATUS_COMPLETED
            - GOAL_STATUS_CANCELED
            - GOAL_STATUS_ARCHIVED
        version:
          type: integer
          format: int64
        createdByUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type:
            - string
            - "null"
          format: date-time
        archivedAt:
          type:
            - string
            - "null"
          format: date-time
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required:
        - companyId
        - title
      properties:
        companyId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 4000
    CreateProjectResponse:
      type: object
      additionalProperties: false
      required:
        - project
      properties:
        project:
          $ref: "#/components/schemas/Project"
    GetProjectResponse:
      type: object
      additionalProperties: false
      required:
        - project
      properties:
        project:
          $ref: "#/components/schemas/Project"
    ListProjectsResponse:
      type: object
      additionalProperties: false
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        nextPageToken:
          type: string
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      required:
        - expectedVersion
      properties:
        expectedVersion:
          type: integer
          format: int64
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 4000
    UpdateProjectResponse:
      type: object
      additionalProperties: false
      required:
        - project
      properties:
        project:
          $ref: "#/components/schemas/Project"
    ChangeProjectStatusResponse:
      type: object
      additionalProperties: false
      required:
        - project
      properties:
        project:
          $ref: "#/components/schemas/Project"
    Project:
      type: object
      additionalProperties: false
      required:
        - tenantId
        - companyId
        - goalId
        - projectId
        - title
        - status
        - version
      properties:
        tenantId:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - PROJECT_STATUS_PLANNED
            - PROJECT_STATUS_ACTIVE
            - PROJECT_STATUS_ON_HOLD
            - PROJECT_STATUS_COMPLETED
            - PROJECT_STATUS_CANCELED
            - PROJECT_STATUS_ARCHIVED
        version:
          type: integer
          format: int64
        createdByUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type:
            - string
            - "null"
          format: date-time
        archivedAt:
          type:
            - string
            - "null"
          format: date-time
...
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          minLength: 1
          examples:
            - internal_error
        message:
          type: string
          minLength: 1
        requestId:
          type:
            - string
            - "null"
        details:
          type:
            - object
            - array
            - string
            - number
            - boolean
            - "null"
  responses:
    BadRequest:
      description: The request was malformed or missing required fields.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The actor lacks the required permission.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with the current resource state.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: An unexpected internal error occurred.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"